split settings into a new file
